requirements.txt
blinker
click
colorama
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
packaging
Werkzeug
flask_cors
flask_jwt_extended
scikit-learn
mysqlclient
dotenv
os